Rename the gazetteer install CLI command from `download` to `install`
